+++ services/ui-src/src/routes/Home/Home.tsx
@@ -29,48 +29,49 @@ export const Home = () => {
 
   return (
     <>
-      {userRole === UserRoles.HELP_DESK ||
-      userRole === UserRoles.APPROVER ||
-      userRole === UserRoles.ADMIN ? (
-        <>
-          <ReadOnly />
-        </>
-      ) : (
-        <>
-          <Collapse in={showBanner}>
-            <Banner bannerData={bannerData} />
-          </Collapse>
-          <BasicPage sx={sx.layout} data-testid="home-view">
-            <Box sx={sx.introTextBox}>
-              <Heading as="h1" sx={sx.headerText}>
-                {intro.header}
-              </Heading>
-              <Text>
-                {intro.body.preLinkText}
-                <Link href={intro.body.linkLocation} isExternal>
-                  {intro.body.linkText}
-                </Link>
-                {intro.body.postLinkText}
-              </Text>
-            </Box>
-            <TemplateCard
-              templateName="MCPAR"
-              verbiage={cards.MCPAR}
-              cardprops={sx.card}
-            />
-            <TemplateCard
-              templateName="MLR"
-              verbiage={cards.MLR}
-              cardprops={sx.card}
-            />
-            <TemplateCard
-              templateName="NAAAR"
-              verbiage={cards.NAAAR}
-              cardprops={sx.card}
-            />
-          </BasicPage>{" "}
-        </>
-      )}
+      <>
+        <Collapse in={showBanner}>
+          <Banner bannerData={bannerData} />
+        </Collapse>
+        <BasicPage sx={sx.layout} data-testid="home-view">
+          {/* show standard view to state users */}
+          {userRole === UserRoles.STATE_USER ||
+          userRole === UserRoles.STATE_REP ? (
+            <>
+              <Box sx={sx.introTextBox}>
+                <Heading as="h1" sx={sx.headerText}>
+                  {intro.header}
+                </Heading>
+                <Text>
+                  {intro.body.preLinkText}
+                  <Link href={intro.body.linkLocation} isExternal>
+                    {intro.body.linkText}
+                  </Link>
+                  {intro.body.postLinkText}
+                </Text>
+              </Box>
+              <TemplateCard
+                templateName="MCPAR"
+                verbiage={cards.MCPAR}
+                cardprops={sx.card}
+              />
+              <TemplateCard
+                templateName="MLR"
+                verbiage={cards.MLR}
+                cardprops={sx.card}
+              />
+              <TemplateCard
+                templateName="NAAAR"
+                verbiage={cards.NAAAR}
+                cardprops={sx.card}
+              />
+            </>
+          ) : (
+            // show read-only view to non-state users
+            <ReadOnly verbiage={verbiage.readOnly} />
+          )}
+        </BasicPage>
+      </>
     </>
   );
 };
